backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
dj-database-url==2.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
psycopg[binary]==3.2.13
sqlparse==0.5.5
tzdata==2025.3
